+++ src/appinfo.h
@@ -1,5 +1,5 @@
-/* wavbreaker - A tool to split a wave file up into multiple waves.
- * Copyright (C) 2002-2003 Timothy Robinson
+/* wavbreaker - A tool to split a wave file up into multiple wave.
+ * Copyright (C) 2022 Thomas Perl
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,11 +16,19 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#ifndef POPUPMESSAGE_H
-#define POPUPMESSAGE_H
+#pragma once
 
-#include <gtk/gtk.h>
+const char *
+appinfo_copyright();
 
-void popupmessage_show(GtkWidget *, const char *, const char *);
+const char *
+appinfo_description();
 
-#endif /* POPUPMESSAGE_H */
+const char *
+appinfo_url();
+
+const char **
+appinfo_authors();
+
+const char *
+appinfo_version();