+++ src/list.h
@@ -1,5 +1,5 @@
-/* wavbreaker - A tool to split a wave file up into multiple waves.
- * Copyright (C) 2002-2003 Timothy Robinson
+/* wavbreaker - A tool to split a wave file up into multiple wave.
+ * Copyright (C) 2022 Thomas Perl
  *
  * This program is free software; you can redistribute it and/or modify
  * it under the terms of the GNU General Public License as published by
@@ -16,11 +16,14 @@
  * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA
  */
 
-#ifndef POPUPMESSAGE_H
-#define POPUPMESSAGE_H
+#pragma once
 
-#include <gtk/gtk.h>
+#include <glib.h>
 
-void popupmessage_show(GtkWidget *, const char *, const char *);
+#include "track_break.h"
 
-#endif /* POPUPMESSAGE_H */
+gboolean
+list_read_file(const char *list_filename, TrackBreakList *list);
+
+gboolean
+list_write_file(const char *list_filename, const char *wav_filename, TrackBreakList *list);